Add files via upload
Description: This program is project milestone 4 towards the term project. It demonstrates building and evaluating a deep neural network to predict bias in news articles. The model's training data was severely limited due to computational power and accurately reflects that. It will be interesting to increase the models training data on a more powerful computer.